requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.27.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
python-dotenv>=1.0.0
tenacity>=8.2.0
datasets>=2.16.0
duckduckgo-search>=4.4.0
sentence-transformers>=2.2.0
numpyro>=0.13.0
jax>=0.4.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
pingouin>=0.5.3
pyyaml>=6.0